requirements.txt
fastapi==0.121.0
uvicorn==0.38.0
python-dotenv==1.1.1
langchain==1.0.3
langchain-core==1.0.3
langchain-openai==1.0.1
langchain-huggingface==1.0.0
sentence-transformers
langchain-community==0.4.1
langchain-text-splitters==1.0.0
faiss-cpu==1.12.0
youtube-transcript-api==1.2.1
chromadb==1.3.2

